extracted interactive session management to interactive.rs
